profiling support to measure latency and throughput

+++ src/runtime_src/xdp/profile/database/static_info/device_info.cpp
@@ -324,7 +324,7 @@ namespace xdp {
 
   void DeviceInfo::addAIECounter(uint32_t i, uint8_t col, uint8_t row,
                                  uint8_t num, uint16_t start, uint16_t end,
-                                 uint8_t reset, uint32_t load, double freq,
+                                 uint8_t reset, uint64_t load, double freq,
                                  const std::string& mod,
                                  const std::string& aieName)
   {
