+++ src/runtime_src/xdp/profile/database/static_info_database.cpp
@@ -1073,7 +1073,7 @@ namespace xdp {
   void VPStaticDatabase::addAIECounter(uint64_t deviceId, uint32_t i,
                                        uint8_t col, uint8_t row, uint8_t num,
                                        uint16_t start, uint16_t end,
-                                       uint8_t reset, uint32_t load,
+                                       uint8_t reset, uint64_t load,
                                        double freq, const std::string& mod,
                                        const std::string& aieName)
   {
